Version 2.0
1) Improved exception handling. Trap different exceptions during
download. Handle according to whether the exception is transient or
permanent either allowing a later on demand reload or ignoring image and
blocking further reload attempts respectively.
2) Detection of network connectivity. During call to Json fetching check
whether there is network connection and if not report to user.
3) Improved resource management during background. Essentially improve
Downloader thread abortion when the application goes into background.
Keep the download queue and restart Download thread upon returning of
application to the foreground.
3) Overall code documentation review